+++ Sources/SwiftUIPanoramaViewer/Original/Extensions.swift
@@ -6,18 +6,133 @@
 //
 
 import Foundation
+import UIKit
 import OSLog
+import GLKit
+import CoreMotion
+import SceneKit
 
 extension Logger {
 
-    fileprivate static var subsystem = "SwiftUIPanoramaViewer" //Bundle(for: PanoramaManager.self).bundleIdentifier!
+    fileprivate static var subsystem = "SwiftUIPanoramaViewer"
 
-    static let panoramaViewer = Logger(subsystem: subsystem, category: "PanoramaViewer")
-    static let sceneKit = Logger(subsystem: subsystem, category: "SceneKit")
+    static let panoramaViewer = Logger(subsystem: Logger.subsystem, category: "PanoramaViewer")
+    static let sceneKit = Logger(subsystem: Logger.subsystem, category: "SceneKit")
 }
 
 extension OSSignposter {
     static let renderer = OSSignposter(subsystem: Logger.subsystem, category: "Renderer")
     static let scene = OSSignposter(subsystem: Logger.subsystem, category: "Scene")
     static let transition = OSSignposter(subsystem: Logger.subsystem, category: "Transition")
+}
+
+extension UIView {
+
+    static func forAutolayout() -> Self {
+        let view = Self()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        return view
+    }
+
+    func add(view: UIView) {
+        view.translatesAutoresizingMaskIntoConstraints = false
+        self.addSubview(view)
+
+        NSLayoutConstraint.activate([
+            view.topAnchor.constraint(equalTo: self.topAnchor),
+            view.leadingAnchor.constraint(equalTo: self.leadingAnchor),
+            view.trailingAnchor.constraint(equalTo: self.trailingAnchor),
+            view.bottomAnchor.constraint(equalTo: self.bottomAnchor)
+        ])
+    }
+}
+
+@MainActor
+extension CMDeviceMotion {
+
+    func orientation() -> SCNVector4 {
+
+        let attitude = self.attitude.quaternion
+        let attitudeQuanternion = GLKQuaternion(quanternion: attitude)
+
+        let result: SCNVector4
+
+        let orientation = UIApplication.shared.connectedScenes
+            .compactMap({ $0 as? UIWindowScene })
+            .first(where: { $0.activationState == .foregroundActive })?
+            .interfaceOrientation
+
+        switch orientation ?? .portrait {
+
+        case .landscapeRight:
+            let cq1 = GLKQuaternionMakeWithAngleAndAxis(.pi/2, 0, 1, 0)
+            let cq2 = GLKQuaternionMakeWithAngleAndAxis(-(.pi/2), 1, 0, 0)
+            var quanternionMultiplier = GLKQuaternionMultiply(cq1, attitudeQuanternion)
+            quanternionMultiplier = GLKQuaternionMultiply(cq2, quanternionMultiplier)
+
+            result = quanternionMultiplier.vector(for: .landscapeRight)
+
+        case .landscapeLeft:
+            let cq1 = GLKQuaternionMakeWithAngleAndAxis(-(.pi/2), 0, 1, 0)
+            let cq2 = GLKQuaternionMakeWithAngleAndAxis(-(.pi/2), 1, 0, 0)
+            var quanternionMultiplier = GLKQuaternionMultiply(cq1, attitudeQuanternion)
+            quanternionMultiplier = GLKQuaternionMultiply(cq2, quanternionMultiplier)
+
+            result = quanternionMultiplier.vector(for: .landscapeLeft)
+
+        case .portraitUpsideDown:
+            let cq1 = GLKQuaternionMakeWithAngleAndAxis(-(.pi/2), 1, 0, 0)
+            let cq2 = GLKQuaternionMakeWithAngleAndAxis(.pi, 0, 0, 1)
+            var quanternionMultiplier = GLKQuaternionMultiply(cq1, attitudeQuanternion)
+            quanternionMultiplier = GLKQuaternionMultiply(cq2, quanternionMultiplier)
+
+            result = quanternionMultiplier.vector(for: .portraitUpsideDown)
+
+        default:
+            let clockwiseQuanternion = GLKQuaternionMakeWithAngleAndAxis(-(.pi/2), 1, 0, 0)
+            let quanternionMultiplier = GLKQuaternionMultiply(clockwiseQuanternion, attitudeQuanternion)
+
+            result = quanternionMultiplier.vector(for: .portrait)
+        }
+        return result
+    }
+}
+
+extension GLKQuaternion {
+
+    init(quanternion: CMQuaternion) {
+        self.init(q: (Float(quanternion.x), Float(quanternion.y), Float(quanternion.z), Float(quanternion.w)))
+    }
+
+    func vector(for orientation: UIInterfaceOrientation) -> SCNVector4 {
+        switch orientation {
+        case .landscapeRight:
+            return SCNVector4(x: -self.y, y: self.x, z: self.z, w: self.w)
+
+        case .landscapeLeft:
+            return SCNVector4(x: self.y, y: -self.x, z: self.z, w: self.w)
+
+        case .portraitUpsideDown:
+            return SCNVector4(x: -self.x, y: -self.y, z: self.z, w: self.w)
+
+        default:
+            return SCNVector4(x: self.x, y: self.y, z: self.z, w: self.w)
+        }
+    }
+}
+
+extension CMMotionManager: @retroactive @unchecked Sendable {}
+
+extension UIColor {
+
+    func render(in size: CGSize, scale: CGFloat = 1) -> UIImage {
+        let format = UIGraphicsImageRendererFormat()
+        format.scale = scale
+
+        let renderer = UIGraphicsImageRenderer(size: size, format: format)
+        return renderer.image { context in
+            self.setFill()
+            context.fill(CGRect(origin: .zero, size: size))
+        }
+    }
 }
